Explicitly use static_cast from size_t to uint8_t in animationLength_get

+++ src/openrct2/scripting/bindings/entity/ScGuest.cpp
@@ -899,7 +899,7 @@ namespace OpenRCT2::Scripting
         }
 
         auto& animationGroup = GetPeepAnimation(peep->SpriteType, peep->ActionSpriteType);
-        return animationGroup.frame_offsets.size();
+        return static_cast<uint8_t>(animationGroup.frame_offsets.size());
     }
 
     ScThought::ScThought(PeepThought backing)
